Added superadmin

+++ sams-app/resources/views/Admin/Admin/index.blade.php
@@ -11,10 +11,12 @@
             </div>
             <div>
                 <div class="float-end">
-                    <button type="button" class="btn btn-success" data-bs-toggle="modal"
-                        data-bs-target="#createModal">+Create</button>
-                    <button class="btn btn-dark float-end" data-bs-toggle="modal" data-bs-target="#archivemodal"><i
-                            class="bi bi-archive"></i></button>
+                    @if (Auth::user()->isAdmin == '2')
+                        <button type="button" class="btn btn-success" data-bs-toggle="modal"
+                            data-bs-target="#createModal">+ Create</button>
+                        <button class="btn btn-dark float-end" data-bs-toggle="modal" data-bs-target="#archivemodal"><i
+                                class="bi bi-archive"></i></button>
+                    @endif
                 </div>
             </div>
         </div>
@@ -34,7 +36,9 @@
                                 <th scope="col" style="text-align: left">ID.</th>
                                 <th scope="col">First Name</th>
                                 <th scope="col">Last Name</th>
-                                <th scope="col">Actions</th>
+                                @if (Auth::user()->isAdmin == '2')
+                                    <th scope="col">Actions</th>
+                                @endif
                             </tr>
                         </thead>
                         <tbody>
@@ -43,24 +47,27 @@
                                     <td class="text-capitalize" style="text-align: left">{{ $admin->id }}</td>
                                     <td class="text-capitalize">{{ $admin->fname }}</td>
                                     <td class="text-capitalize">{{ $admin->lname }}</td>
-                                    <td>
-                                        <div class="d-flex gap-1 align-items-center">
-                                            <button type="button" class="btn btn-success btn-sm" data-bs-toggle="modal"
-                                                data-bs-target="#updateModal_{{ $admin->id }}"><i
-                                                    class="bi bi-pencil"></i></button>
-                                            {{-- Modal Update --}}
-                                            @include('Admin.Admin.Modal.update-modal', [
-                                                'Admin' => $admin,
-                                            ])
-                                            <form method="POST" action="{{ route('admin.destroy', $admin->id) }}">
-                                                @csrf
-                                                @method('DELETE')
-                                                <button type="submit" class="btn bg-danger btn-sm"
-                                                    onclick="deleteConfirmation(event, this)"><i
-                                                        class="bi bi-trash-fill"></i></button>
-                                            </form>
-                                        </div>
-                                    </td>
+                                    @if (Auth::user()->isAdmin == '2')
+                                        <td>
+                                            <div class="d-flex gap-1 align-items-center">
+                                                <button type="button" class="btn btn-success btn-sm"
+                                                    data-bs-toggle="modal"
+                                                    data-bs-target="#updateModal_{{ $admin->id }}"><i
+                                                        class="bi bi-pencil"></i></button>
+                                                {{-- modal here --}}
+                                                @include('Admin.Admin.Modal.update-modal', [
+                                                    'Admin' => $admin,
+                                                ])
+                                                <form method="POST" action="{{ route('admin.destroy', $admin->id) }}">
+                                                    @csrf
+                                                    @method('DELETE')
+                                                    <button type="submit" class="btn bg-danger btn-sm"
+                                                        onclick="deleteConfirmation(event, this)"><i
+                                                            class="bi bi-trash-fill"></i></button>
+                                                </form>
+                                            </div>
+                                        </td>
+                                    @endif
                                 </tr>
                             @endforeach
                         </tbody>
